Add top-level debug handler (prep'ing for global debug module)
Add support for polling CME time (start, stop, etc).  The Action module is responsible for dispatching a time request periodically and the Store receives the dispatched response to update the CME time field in the config.  Default polling interval is 1 second.

Support the more generic version of config (GET, POST through the API) to get or set whole or portions of the config object tree.