Add top-level debugging handler
Add a Store reference to retrieve the client-side CME configuration object.  This probably should go the other way 'round (Store retrieves CME config from API), so this will likely change.  But for now, the API supports gets/sets from/to the config tree in a flexible fashion so that whole or portions of the config tree may be updated or read.

Adds polling support.  The API poll function takes a url (fragment) from which it sends a GET request.  Success and failure callbacks are called when responses are received (or not).  No further response processing is done - the success callback is expected to handle the response appropriately.

Supports more generic 'config' requests to GET or POST updates from or to the CME configuration.