Updates the API status - channel - sensors/controls hierarchy and class implementations.
Essentially, the hw status object is read from memcached to get the latest values.  Then the hw object hierarchy is used to create a supplemented object hierarchy with the hardware data plus additional attributes like names, descriptions, etc., that may want to be set by the user.  The resulting object is returned to the user via the API routes.