Major UI restructuring.  Moved the polling out of top-level (Actions.jsx) to better allow components to be responsible for their own polling rates, handling, etc.
So now, for example, each channel panel present is responsible to poll for its own data.  This alleviates a huge bottleneck when a channel history is being polled (it can take some time to retrieve for longer histories) and allows the other channels to continue with their own updates.

*** WARNING ***
This is still fairly untested, but seemed to "work for me" so I committed it!

There are some refinements (and fixes) I'm sure that will be necessary.